YUK-10233 Server Messaging-Improve Capbank and Zone widget on the zone detail page to include IVVC server analysis information.
Updated the subbus ID to be an int (as the client expects) instead of a long.

+++ yukon-server/Message/IVVCAnalysisMessage.cpp
@@ -10,7 +10,7 @@ namespace Messaging     {
 namespace CapControl    {
 
 
-IVVCAnalysisMessage::IVVCAnalysisMessage( const long      subbusId,
+IVVCAnalysisMessage::IVVCAnalysisMessage( const int       subbusId,
                                           const int       scenarioId,
                                           const CtiTime & timestamp )
     : _subbusId( subbusId ),
@@ -97,7 +97,7 @@ IVVCAnalysisMessage * IVVCAnalysisMessage::createExceedMaxKVarMessage( const int
 
 void IVVCAnalysisMessage::streamInto( cms::StreamMessage & message ) const
 {
-    message.writeLong( _subbusId );
+    message.writeInt( _subbusId );
     message.writeLong( _timestamp );
     message.writeInt( _scenarioId );
     message.writeInt( _intData.size() );
